+++ Changelog.md
@@ -2,6 +2,7 @@ HibernationFixup Changelog
 ============================
 #### v1.3.4
 - Improve auto-hibernate feature: correct next wake time disregarding the current sleep phase.
+- Added constants for 11.0 support.
 
 #### v1.3.3
 - Improve auto-hibernate feature: support standby and autopoweroff separately (with respective delay). Immediate hibernate is possible with zero delay.